Projeto/RoveriBack_final/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2.0
Pillow>=9.0
django-cors-headers>=3.14
psycopg2-binary>=2.9
channels>=4.0
daphne>=4.0
channels-redis>=4.0
djangorestframework-simplejwt>=5.2.0
django-redis>=5.2.0
